+++ src/allmydata/test/test_storage.py
@@ -35,24 +35,8 @@ from allmydata.test.no_network import NoNetworkServer
 from allmydata.storage_client import (
     _StorageServer,
 )
+from .common_py3 import FakeCanary
 
-class Marker(object):
-    pass
-
-class FakeCanary(object):
-    def __init__(self, ignore_disconnectors=False):
-        self.ignore = ignore_disconnectors
-        self.disconnectors = {}
-    def notifyOnDisconnect(self, f, *args, **kwargs):
-        if self.ignore:
-            return
-        m = Marker()
-        self.disconnectors[m] = (f, args, kwargs)
-        return m
-    def dontNotifyOnDisconnect(self, marker):
-        if self.ignore:
-            return
-        del self.disconnectors[marker]
 
 class FakeStatsProvider(object):
     def count(self, name, delta=1):
